Use nom-rule on crates.io
