+++ common/ast/Cargo.toml
@@ -16,13 +16,13 @@ common-exception = { path = "../exception" }
 common-functions = { path = "../functions" }
 
 # Github dependencies
-nom-rule = { git = "https://github.com/andylokandy/nom-rule" }
 sqlparser = { git = "https://github.com/datafuse-extras/sqlparser-rs", rev = "c33837e" }
 
 # Crates.io dependencies
 async-trait = "0.1.52"
 logos = "0.12"
 nom = "7"
+nom-rule = "0.1"
 once_cell = "1.9.0"
 thiserror = "1.0.30"
